move stringToTa to proper util
