+++ packages/shared/src/logic/utils.ts
@@ -104,49 +104,3 @@ export function normalizeUrbitColor(color: string): string {
   const lengthAdjustedColor = _.padStart(colorString, 6, '0');
   return `#${lengthAdjustedColor}`;
 }
-
-//  encode the string into @ta-safe format, using logic from +wood.
-//  for example, 'some Chars!' becomes '~.some.~43.hars~21.'
-//  this is equivalent to (scot %t string), and is url-safe encoding for
-//  arbitrary strings.
-//
-//  TODO  should probably go into aura-js
-export function stringToTa(string: string) {
-  let out = '';
-  for (let i = 0; i < string.length; i += 1) {
-    const char = string[i];
-    let add = '';
-    switch (char) {
-      case ' ':
-        add = '.';
-        break;
-      case '.':
-        add = '~.';
-        break;
-      case '~':
-        add = '~~';
-        break;
-      default: {
-        const codePoint = string.codePointAt(i);
-        if (!codePoint) break;
-        //  js strings are encoded in UTF-16, so 16 bits per character.
-        //  codePointAt() reads a _codepoint_ at a character index, and may
-        //  consume up to two js string characters to do so, in the case of
-        //  16 bit surrogate pseudo-characters. here we detect that case, so
-        //  we can advance the cursor to skip past the additional character.
-        if (codePoint > 0xffff) i += 1;
-        if (
-          (codePoint >= 97 && codePoint <= 122) || // a-z
-          (codePoint >= 48 && codePoint <= 57) || // 0-9
-          char === '-'
-        ) {
-          add = char;
-        } else {
-          add = `~${codePoint.toString(16)}.`;
-        }
-      }
-    }
-    out += add;
-  }
-  return `~~${out}`;
-}
